Migrate from deprecated usages (#5)
Sui SDK is going through a refactor. Trying to keep up.

Note that this revealed a behavioral change in signing and verifying
personal messages, which is currently failing CI. The bug is tracked by
[sc-1336].

[UPDATE] - add support for `wrapBcs` on `signPersonalMessage` call,
which defaults to `true` to match Sui SDK behavior. (The CI failure will
be resolved once updated services are deployed).