Configure coverage and add codecov (#44)
Had another open PR from a public fork, but github checks are having
auth issues so it might be easier to configure from Shinami org.
Copy-pasting everything from other PR.

[sc-1249]

Full demo of codecov changes is here:
https://github.com/ivanpeng/shinami-typescript-sdk/pull/2

This needs a few things to be enabled on repo side:
- codecov app needs to be enabled by repo admin
- codecov token needs to be set as a secret in repo settings

After this PR is merged to main, codecov will auto-start commenting on
the subsequent PRs with coverage details. We can also enable the codecov
checks on the PR if we want