Fix compatibility with latest Tanstack Query (#51)
[This](https://github.com/TanStack/query/pull/9615/files#diff-988cd43951b30185641ccb757eefd39c6e67027d7ccab1f445117b4c0401dc8bR1102)
was the specific breaking change introduced in tanstack v5.89.0. It
introduced a second mutation context parameter needed in
MutationFunction. ~~The fix presented just adds an empty context as the
second param to the mutation function calls since we only appear to use
the params, not the context.~~

The mutation functions we provide don't make use of the context, so no
need to update the dependency version. However, we modify our code to be
compilable against both versions.

Also update openrpc client to address [sc-6497]

---------

Co-authored-by: Hao Xia <hao@shinami.com>